refactored Course Controls dashboard to reusable component, and abstracted the section/course edit screen to it's own component (SectionIndex)
